scripts/post-fs-data: introduce susfs-assisted global whiteouts
Having whiteouts is a very useful feature that is built onto KernelSU
by utillizing overlayfs. but theres an issue when you have susfs auto,
susfs tends to unmount the whiteouts, which makes the feature useless,
especially when the default on KernelSU is umount by default.

This change restores that and does it generically via overlayfs. It does
not matter if you have Overlayfs or magic mount KernelSU, as this method
globally mounts the whiteouts on its own via OverlayFS.

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>