scripts/service: Rework sus_su checks
Since susfs v1.5.3 introduced the "show enabled_features" in the userspace command, we can now use this to check if the SU is supported or not.
Else we will fallback to the old method of checking if sus_su command outputs nothing (which is supported).
also, added a check when sus_su config is 0.