module: scripts/customize: use internet connection to install/update susfs binaries
since we're using local files to install the susfs binaries,
the problem with this one is that if we keep on adding local susfs binaries to the zip file, the zip file will be inflate or gets larger and lareger.
so we need to use the internet connection to download the latest files from the internet.
If the internet connection or the binaries of that version is not available, we can use the local files instead.

the source of the susfs binaries is from the following link:
https://github.com/sidex15/susfs4ksu-binaries