Do not show sus_su is not available message when the kernel is NON-GKI
SUS SU is deprecated and will not support NON-GKI kernels starting v1.5.4
This commit is also affects SUSFS v1.5.3

Why?
Because NON-GKI kernels have so-called "KernelSU Manual Hooks". It's in the documentation of KernelSU
What SUS_SU 2 is doing is just do the KernelSU Manual Hooks method of NON-GKI for GKI kernels.
Which prevents it from using the kprobes hooking method which is detected by most detectors out there.

Basically:
Manual hooks of NON-GKI > Kprobe hooks

That's why SUS_SU is not needed in NON-GKI kernels.